修改登录返回值，修改上传报告bug，增加报告评分api

+++ src/main/java/com/project/etsapi/entity/Report.java
@@ -12,13 +12,13 @@ public class Report {
     private String project_name;
     private String student_ID;
     private String report_name;
-    private Double score;
+    private Integer score;
     private String submit_time;
     private String correct_time;
 
     public Report(String course_ID, String project_name,
                   String student_ID, String report_name,
-                  Double score, String submit_time, String correct_time) {
+                  Integer score, String submit_time, String correct_time) {
         this.course_ID = course_ID;
         this.project_name = project_name;
         this.student_ID = student_ID;
@@ -74,11 +74,11 @@ public class Report {
         this.report_name = report_name;
     }
 
-    public Double getScore() {
+    public Integer getScore() {
         return score;
     }
 
-    public void setScore(Double score) {
+    public void setScore(Integer score) {
         this.score = score;
     }
